requirements.txt
ply~=3.11
google-generativeai~=0.8.5
python-dotenv~=1.1.1
protobuf~=5.29.5